requirements.txt
streamlit
pandas
requests
matplotlib
plotly
google-generativeai
SpeechRecognition
streamlit-webrtc

